Introduce encoding adapters
This introduces an Adapter interface that encodings can optionally use
to delegate the (de)serialization of those types to another type. This
is useful, for example, supporting an aggressive Thrift to Protobuf
migration; Thrift users wanting Protobuf on the wire would not need to
change any code.

A no-op implementation is introduced to make generated code logic,
more readable.