Introduce `direct` peer chooser
This introduces a new peer chooser, `direct`, that returns the peer
identified by `transport.Request#ShardKey`. This supports directly
addressable peers across transports.

The `direct` chooser is not a `peer.List` as it only holds onto peers
for as long as it takes for a call to succeed. Therefore, it has no
`Update` method.

The configuration structs are added for forwards
compatibility. Service owners config can register the
`yarpcconfig.PeerChooserSpec` and configure the peer chooser as
follows:

 ```
 outbounds:
   destination-service:
     grpc:
	direct: {}
 ```